+++ mod-source-record-storage-server/src/test/java/org/folio/verticle/consumers/DataImportConsumersVerticleTest.java
@@ -201,7 +201,7 @@ public class DataImportConsumersVerticleTest extends AbstractLBServiceTest {
     var value = DI_SRS_MARC_BIB_RECORD_MODIFIED_READY_FOR_POST_PROCESSING.value();
     String observeTopic = getTopicName(value);
     List<KeyValue<String, String>> observedRecords = cluster.observe(ObserveKeyValues.on(observeTopic, 1)
-      .observeFor(30, TimeUnit.SECONDS)
+      .observeFor(50, TimeUnit.SECONDS)
       .build());
 
     Event obtainedEvent = Json.decodeValue(observedRecords.get(0).getValue(), Event.class);